+++ components/studio/api/urls.py
@@ -1,7 +1,8 @@
 from django.conf.urls import include
 from django.urls import path
 import rest_framework.routers as drfrouters
-from .views import ModelList, ModelLogList, MetadataList, ReportList, ReportGeneratorList, ProjectList, MembersList, DatasetList
+from .views import ModelList, ModelLogList, MetadataList, ProjectList, MembersList, ObjectTypeList
+from .views import AppInstanceList, FlavorsList, EnvironmentList, S3List, MLflowList, ResourceList
 from .public_views import get_studio_settings
 from rest_framework.authtoken.views import obtain_auth_token
 from rest_framework_nested import routers
@@ -12,19 +13,20 @@ router_drf = drfrouters.DefaultRouter()
 
 router = routers.SimpleRouter()
 
-router.register(r'reports', ReportList, basename='report')
-router.register(r'generators', ReportGeneratorList, basename='report_generator')
 router.register(r'projects', ProjectList, basename='project')
 
 models_router = routers.NestedSimpleRouter(router, r'projects', lookup='project')
 models_router.register(r'models', ModelList, basename='model')
+models_router.register(r'objecttype', ObjectTypeList, basename='objecttype')
 models_router.register(r'members', MembersList, basename='members')
-models_router.register(r'dataset', DatasetList, basename='dataset')
+models_router.register(r'resources', ResourceList, basename='resources')
+models_router.register(r'appinstances', AppInstanceList, basename='appinstances')
+models_router.register(r'flavors', FlavorsList, basename='flavors')
+models_router.register(r'environments', EnvironmentList, basename='environment')
+models_router.register(r's3', S3List, basename='s3')
+models_router.register(r'mlflow', MLflowList, basename='mlflow')
 models_router.register(r'modellogs', ModelLogList, basename='modellog')
 models_router.register(r'metadata', MetadataList, basename='metadata')
-
-# router.register(r'deploymentInstances', DeploymentInstanceList, basename='deploymentInstance')
-# router.register(r'deploymentDefinitions', DeploymentDefinitionList, basename='deploymentDefinition')
 
 urlpatterns = [
     path('', include(router_drf.urls)),