+++ components/studio/scripts/run_worker.sh
@@ -16,4 +16,5 @@ cd ..
 
 sleep 1
 
-watchmedo auto-restart -R --patterns="*.py" -- celery -A studio worker -l info --beat --scheduler django
+#  watchmedo auto-restart -R --patterns="*.py" -- 
+celery -A studio worker -l info --beat --scheduler django