+++ docs/releasenotes.md
@@ -5,6 +5,7 @@
 ### New functionality
 - Studio Administration module to monitor and manage resources as an alternative to Django Admin
 - Brand new STACKn UI
+- Automated scaling of model deployments
 
 ### Bug fixes